docs:  make Doxyfile independently useful...
The place-holder "tags" in Doxyfile are no longer needed thanks
to the introduction of `doxygen_config.py` module that now gives us
an easy interface to the Doxygen settings during the doc-build process.
The Doxyfile settings pertinent to the doc-build are now set (overridden)
during the doc-build without altering the original Doxyfile.